+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+Nothing here yet.
+
+## [v1.1.0](https://github.com/mnemonikr/libsla/tree/v1.1.0)
+
 ### Added
 
 * Support for serializing `libsla` structures using `serde`